+++ app/coin-swiper.css
