+++ app/sharedinterfaces.ts
@@ -1,0 +1,3 @@
+export interface Coin {
+  value: string;
+}
