--- components/Currencies.tsx
@@ -1,66 +1,0 @@
-
-export const CurrencyArray = [
-  "btc",
-  "eth",
-  "ltc",
-  "bch",
-  "bnb",
-  "eos",
-  "xrp",
-  "xlm",
-  "link",
-  "dot",
-  "yfi",
-  "usd",
-  "aed",
-  "ars",
-  "aud",
-  "bdt",
-  "bhd",
-  "bmd",
-  "brl",
-  "cad",
-  "chf",
-  "clp",
-  "cny",
-  "czk",
-  "dkk",
-  "eur",
-  "gbp",
-  "gel",
-  "hkd",
-  "huf",
-  "idr",
-  "ils",
-  "inr",
-  "jpy",
-  "krw",
-  "kwd",
-  "lkr",
-  "mmk",
-  "mxn",
-  "myr",
-  "ngn",
-  "nok",
-  "nzd",
-  "php",
-  "pkr",
-  "pln",
-  "rub",
-  "sar",
-  "sek",
-  "sgd",
-  "thb",
-  "try",
-  "twd",
-  "uah",
-  "vef",
-  "vnd",
-  "zar",
-  "xdr",
-  "xag",
-  "xau",
-  "bits",
-  "sats"
-]
-
