--- components/Every_nth.tsx
@@ -1,1 +1,0 @@
-export const every_nth = (arr, nth) => arr?.filter((e, i) => i % nth === nth - 1);
