--- components/UnixTimeConverter.tsx
@@ -1,7 +1,0 @@
-export function convertUnixToDate(time) {
-    const date = new Date(time);
-    const year = date.getFullYear();
-    const month = date.getMonth() + 1;
-    const day = date.getDate();
-    return `${month}/${day}`;
-  }
